Add the implementation of matmul_t_v2.cu
matmul_t_v2.cu is based on matmul_t_v1.cu, and it uses a three stage
pipeline to overlap computation and global to shared memory data load.
The performance improvement relative to matmul_t_v1.cu is relatively
low, presumably because the performance bottleneck now is actually the
shared to register data loading process.

The TFLOPS is 110.34 / 222.11 = 49.67% of cuBLAS. The blockdim is
(16, 16). The block tile shape is (256, 128) for tileC, (256, 16) for
tileA and (16, 128) for tileB. The warp tile shape is
(64, 64), and the grid dim is (16, 32).

+++ task-1/src/transpose_mat.cu
@@ -11,7 +11,7 @@ void TransposeMatFP32(int m, int n, float *mat) {
 void TransposeMatFP16(int m, int n, half *mat) {
     for (int i = 0; i < m; i++) {
         for (int j = i + 1; j < n; j++) {
-            float tmp = mat[i * n + j];
+            half tmp = mat[i * n + j];
             mat[i * n + j] = mat[j * n + i];
             mat[j * n + i] = tmp;
         }
